+++ src/TObj/TObj_Application.cxx
@@ -126,7 +126,7 @@ Standard_Boolean TObj_Application::LoadDocument
   myIsError = aStatus != CDF_RS_OK;
   if (myIsError)
   {
-    switch (aStatus)
+    switch ((int)aStatus)
     {
     case CDF_RS_UnknownDocument:
       ErrorMessage (Message_Msg("TObj_Appl_RUnknownDocument") << aPath);
